Separate types, config expire values
